+++ desktop/panel/database.test.js
@@ -85,76 +85,6 @@ for (const subprocess of RUNNERS) {
     );
   }
 
-  describe('elasticsearch testdata/documents tests', () => {
-    const tests = [
-      {
-        query: '',
-        range: null,
-        results: 4,
-      },
-      {
-        query: 'pageCount:>0',
-        range: null,
-        results: 3,
-      },
-      {
-        query: 'pageCount:<0',
-        range: null,
-        results: 0,
-      },
-      {
-        query: 'pageCount:>0',
-        range: {
-          field: 'publishedDate.$date',
-          rangeType: 'absolute',
-          begin_date: new Date('2008-01-01'),
-          end_date: new Date('2010-01-01'),
-        },
-        results: 2,
-      },
-    ];
-
-    for (const testcase of tests) {
-      test(`runs ${JSON.stringify(testcase)} query`, async () => {
-        if (process.platform !== 'linux') {
-          return;
-        }
-
-        const connectors = [
-          new DatabaseConnectorInfo({
-            type: 'elasticsearch',
-          }),
-        ];
-        const dp = new DatabasePanelInfo();
-        dp.database.connectorId = connectors[0].id;
-        dp.database.table = 'test';
-        dp.database.range = testcase.range;
-        dp.content = testcase.query;
-
-        let finished = false;
-        const panels = [dp];
-        await withSavedPanels(
-          panels,
-          async (project) => {
-            const panelValueBuffer = fs.readFileSync(
-              getProjectResultsFile(project.projectName) + dp.id
-            );
-
-            const v = JSON.parse(panelValueBuffer.toString());
-            expect(v.length).toBe(testcase.results);
-
-            finished = true;
-          },
-          { evalPanels: true, connectors, subprocessName: subprocess }
-        );
-
-        if (!finished) {
-          throw new Error('Callback did not finish');
-        }
-      });
-    }
-  });
-
   describe('influx testdata/influx tests', () => {
     const tests = [
       {
